Remove Semester History File
This file is no longer necessary with the new workflow: The alltime history takes care of preventing a paper from being selected a second time. The other use in the previous code base was to re-generate the website, but this no done "locally", i.e. without knowing the entire semester history (by appending to `papers.md`), so this use case is also gone.